Extend create() and add() to support accessors
It is now possible to declare accessors (getters/setters) using the
property list extension of Type.create() and Type.add(). Doing so will
make all the property names in the list accessors and result in the data
being "hidden" in an off-result data object, which can optionally be
referenced withing the result to make the data directly accessible.